+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * fix: Mixed & untyped property was not previously working.
 * feat(`DataCollector`): Add target type hint information.
 * fix(`ObjectToObjectMetadataFactory`): Fix scalar type logic.
+* test: Fix psalm errors.
 
 ## 0.7.0